requirements.txt
mcp>=0.1.0
httpx>=0.25.0
anthropic
python-dotenv
openai